更改int

+++ 1-2.py
@@ -1,5 +1,5 @@
 # Read the numbers b and h like this:
 b = int(input())
 h = int(input())
-print(int(b*h*1/2))
+print(b*h*1/2)
 # Print the result with
